Combine duplicate test files ending with -1 (#56031)
